requirements.txt
Django==4.2
djangorestframework==3.14
celery==5.3
redis==5.0
psycopg2-binary==2.9
